%u negativ dont work, %p not work, %s 1 case segfault

+++ libft/libft.h
@@ -6,13 +6,14 @@
 /*   By: lcadinot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 12:26:59 by lcadinot          #+#    #+#             */
-/*   Updated: 2022/11/20 16:39:50 by lcadinot         ###   ########.fr       */
+/*   Updated: 2022/11/22 18:48:10 by lcadinot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
+# include <limits.h>
 # include <stdarg.h>
 # include <unistd.h>
 # include <stdlib.h>
@@ -55,7 +56,7 @@ void		ft_putnbr_fd(int n, int fd);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
 char		*ft_strjoin(const char *s1, const char *s2);
 char		*ft_strtrim(char const *s1, char const *set);
-char		*ft_itoa(long int n);
+char		*ft_itoa(int n);
 char		**ft_split(char const *s, char c);
 char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void		ft_striteri(char *s, void (*f)(unsigned int, char *));
